refactor: HttpResponse 로직 정리
