+++ tomcat/src/main/java/org/apache/controller/Controller.java
@@ -1,7 +1,7 @@
 package org.apache.controller;
 
-import org.apache.http.HttpResponse;
 import org.apache.http.request.HttpRequest;
+import org.apache.http.response.HttpResponse;
 
 public interface Controller {
