+++ tomcat/src/main/java/org/apache/http/response/HttpResponse.java
@@ -1,4 +1,4 @@
-package org.apache.http;
+package org.apache.http.response;
 
 import java.io.IOException;
 import java.io.OutputStream;
@@ -9,6 +9,7 @@ import java.util.List;
 import java.util.Map;
 import org.apache.exception.RequestProcessingException;
 import org.apache.exception.SocketWriteException;
+import org.apache.http.Cookie;
 import org.apache.http.value.HttpHeader;
 import org.apache.http.value.HttpVersion;
 import org.apache.http.value.StatusCode;
@@ -26,12 +27,11 @@ public class HttpResponse {
     }
 
     public void writeMessage(OutputStream outputStream) {
-        String message = getMessage();
         try {
+            String message = body == null ? makeMassageWhenEmptyBody() : makeMassageWhenBody();
             outputStream.write(message.getBytes(StandardCharsets.UTF_8));
             outputStream.flush();
         } catch (IOException e) {
-            e.printStackTrace();
             throw new SocketWriteException("소켓에 데이터를 쓰는중 오류가 발생했습니다.");
         }
     }
@@ -61,6 +61,26 @@ public class HttpResponse {
         setHeader(HttpHeader.LOCATION.getValue(), location);
     }
 
+    private String makeMassageWhenBody() {
+        validateCanMakeMessage();
+        List<String> messageLines = new ArrayList<>();
+        messageLines.add(makeStartLine());
+        messageLines.addAll(makeCustomHeaderLine());
+        messageLines.addAll(makeSetCookieHeaderLine());
+        messageLines.add(makeContentLengthHeaderLine());
+        return String.join("\r\n", messageLines) + "\r\n" + "\r\n" + body;
+    }
+
+    private String makeMassageWhenEmptyBody() {
+        validateCanMakeMessage();
+        List<String> messageLines = new ArrayList<>();
+        messageLines.add(makeStartLine());
+        messageLines.addAll(makeCustomHeaderLine());
+        messageLines.addAll(makeSetCookieHeaderLine());
+        return String.join("\r\n", messageLines) + "\r\n" + "\r\n";
+    }
+
+
     private String makeStartLine() {
         return String.format("%s %s %s",
                 httpVersion.getValue(),
@@ -68,56 +88,37 @@ public class HttpResponse {
                 statusCode.getMessage());
     }
 
-    private String makeHeaderLines() {
+    private List<String> makeCustomHeaderLine() {
         List<String> headerLines = new ArrayList<>();
-        addCustomHeaderLine(headerLines);
-        addSetCookieHeaderLine(headerLines);
-        addContentLengthHeaderLine(headerLines);
-        return String.join("\r\n", headerLines);
-    }
-
-    private void addCustomHeaderLine(List<String> headerLines) {
         List<String> customHeaderKeys = headers.keySet().stream().toList();
         for (String key : customHeaderKeys) {
             String value = headers.get(key);
             headerLines.add(key + ": " + value);
         }
+        return headerLines;
     }
 
-    private void addSetCookieHeaderLine(List<String> headerLines) {
+    private List<String> makeSetCookieHeaderLine() {
+        List<String> cookieLines = new ArrayList<>();
         if (cookies.isEmpty()) {
-            return;
+            return cookieLines;
         }
+
         for (Cookie cookie : cookies) {
-            String setCookieHeader = String.format("%s: %s",
-                    HttpHeader.SET_COOKIE.getValue(),
-                    cookie.makeCookieLine());
-            headerLines.add(setCookieHeader);
+            cookieLines.add(HttpHeader.SET_COOKIE.getValue() + ": " + cookie.makeCookieLine());
         }
+        return cookieLines;
     }
 
-    private void addContentLengthHeaderLine(List<String> headerLines) {
-        if (body == null) {
-            return;
-        }
-        String contentLengthHeader = String.format("%s: %s",
+    private String makeContentLengthHeaderLine() {
+        return String.format("%s: %s",
                 HttpHeader.CONTENT_LENGTH.getValue(),
                 body.getBytes(StandardCharsets.UTF_8).length);
-        headerLines.add(contentLengthHeader);
     }
 
     private void validateCanMakeMessage() {
         if (httpVersion == null || statusCode == null) {
             throw new RequestProcessingException("주요 응답 필드가 비어있어 응답 메세지를 생성할 수 없습니다.");
         }
-    }
-
-    private String getMessage() {
-        validateCanMakeMessage();
-        String startAndHeader = String.join("\r\n", makeStartLine(), makeHeaderLines()) + "\r\n";
-        if (body != null && !body.isEmpty()) {
-            return startAndHeader + "\r\n" + body;
-        }
-        return startAndHeader;
     }
 }
